+++ src/__tests__/run.test.ts
@@ -47,6 +47,95 @@ describe('Config without a provided command', () => {
       pulumiVersion: '^2',
     });
   });
+
+  it('should read version from pulumi-version-file', async () => {
+    jest.mock('fs', () => ({
+      ...jest.requireActual('fs'),
+      readFileSync: jest.fn((path: string) => {
+        expect(path).toEqual('.pulumi.version');
+        return '3.121.0';
+      }),
+      existsSync: jest.fn(() => {
+        return true;
+      }),
+    }));
+    jest.mock('@actions/core', () => ({
+      getInput: jest.fn((name: string) => {
+        switch (name) {
+          case 'pulumi-version-file':
+            return '.pulumi.version';
+          case 'pulumi-version':
+            return undefined;
+        }
+        return installConfig[name];
+      }),
+    }));
+
+    const { makeInstallationConfig } = require('../config');
+    const conf = makeInstallationConfig();
+    expect(conf.success).toBeTruthy();
+    expect(conf.value).toEqual({
+      command: undefined,
+      pulumiVersion: '3.121.0',
+    });
+  });
+
+  it('should fail if pulumi-version-file does not exist', async () => {
+    jest.mock('fs', () => ({
+      ...jest.requireActual('fs'),
+      readFileSync: jest.fn((path: string) => {
+        expect(path).toEqual('.pulumi.version');
+        return '3.121.0';
+      }),
+      existsSync: jest.fn(() => {
+        return false;
+      }),
+    }));
+    jest.mock('@actions/core', () => ({
+      getInput: jest.fn((name: string) => {
+        switch (name) {
+          case 'pulumi-version-file':
+            return '.pulumi.version';
+          case 'pulumi-version':
+            return undefined;
+        }
+        return installConfig[name];
+      }),
+    }));
+
+    const { makeInstallationConfig } = require('../config');
+    expect(() => {
+      makeInstallationConfig();
+    }).toThrow(/pulumi-version-file '\.pulumi\.version' does not exist/);
+  });
+
+  it('should fail if pulumi-version-file and pulumi-version are both provided', async () => {
+    jest.mock('fs', () => ({
+      ...jest.requireActual('fs'),
+      readFileSync: jest.fn((path: string) => {
+        expect(path).toEqual('.pulumi.version');
+        return '3.121.0';
+      }),
+      existsSync: jest.fn(() => {
+        return false;
+      }),
+    }));
+    jest.mock('@actions/core', () => ({
+      getInput: jest.fn((name: string) => {
+        if (name === 'pulumi-version-file') {
+          return '.pulumi.version';
+        }
+        return installConfig[name];
+      }),
+    }));
+
+    const { makeInstallationConfig } = require('../config');
+    expect(() => {
+      makeInstallationConfig();
+    }).toThrow(
+      /Only one of 'pulumi-version' or 'pulumi-version-file' should be provided, got both/,
+    );
+  });
 });
 
 describe('main.login', () => {